+++ src/main/java/com/easypost/exception/API/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Custom API exception classes for the EasyPost API.
+ *
+ * @author EasyPost developers
+ * @version 6.0
+ * @see <a href="https://www.easypost.com/docs">EasyPost API documentation</a>
+ * @since 6.0
+ */
+package com.easypost.exception.API;