+++ src/main/java/com/easypost/exception/EasyPostException.java
@@ -8,50 +8,100 @@
 
 package com.easypost.exception;
 
+import com.easypost.model.Error;
+
+import java.util.List;
+
 public class EasyPostException extends Exception {
 
     private static final long serialVersionUID = 1L;
-    private final String param;
+    private final String code;
+    private final Integer statusCode;
+    private final String message;
+    private final List<Error> errors;
 
     /**
-     * Get the parameter name.
-     *
-     * @return the parameter name
-     */
-    public String getParam() {
-        return param;
-    }
-
-    /**
-     * Constructor.
+     * EasyPostException constructor.
      *
      * @param message the exception message
      */
     public EasyPostException(final String message) {
-        super(message, null);
-        this.param = null;
+        this(message, null);
     }
 
     /**
-     * Constructor.
+     * EasyPostException constructor.
      *
      * @param message the exception message
      * @param ex      the exception cause
      */
     public EasyPostException(final String message, final Throwable ex) {
-        super(message, ex);
-        this.param = null;
+        this(message, null, null, null, ex);
     }
 
     /**
-     * Constructor.
+     * EasyPostException constructor.
      *
-     * @param message the exception message
-     * @param param   the parameter name
-     * @param ex      the exception cause
+     * @param message    the exception message
+     * @param code       the exception code
+     * @param statusCode the exception status code
+     * @param errors     the errors array
      */
-    public EasyPostException(final String message, final String param, final Throwable ex) {
+    public EasyPostException(final String message, final String code, final int statusCode, List<Error> errors) {
+        this(message, code, statusCode, errors, null);
+    }
+
+    /**
+     * EasyPostException constructor.
+     *
+     * @param message    the exception message
+     * @param code       the exception code
+     * @param statusCode the exception status code
+     * @param errors     the errors array
+     * @param ex         the exception cause
+     */
+    public EasyPostException(final String message, final String code, final Integer statusCode,
+        final List<Error> errors, final Throwable ex) {
         super(message, ex);
-        this.param = param;
+        this.code = code;
+        this.statusCode = statusCode;
+        this.message = message;
+        this.errors = errors;
+    }
+
+    /**
+     * Get status code of the error object.
+     * 
+     * @return statusCode the status code of the error object
+     */
+    public Integer getStatusCode() {
+        return statusCode;
+    }
+
+    /**
+     * Get code of the error object.
+     *
+     * @return code the code of the error object
+     */
+    public String getCode() {
+        return code;
+    }
+
+    /**
+     * Get message of the error object.
+     *
+     * @return message the message of the error object
+     */
+    public String getMessage() {
+        return message;
+    }
+
+    /**
+     * Get errors array of the exception.
+     *
+     * @return errors of the exception
+     */
+    public List<Error> getErrors() {
+        return errors;
     }
 }