+++ src/main/java/com/easypost/exception/General/ExternalApiError.java
@@ -1,0 +1,14 @@
+package com.easypost.exception.General;
+
+import com.easypost.exception.EasyPostException;
+
+public class ExternalApiError extends EasyPostException {
+    /**
+     * ExternalApiError constructor.
+     *
+     * @param message the exception message
+     */
+    public ExternalApiError(final String message) {
+        super(message);
+    }
+}