+++ src/main/java/com/easypost/exception/General/InvalidObjectError.java
@@ -1,0 +1,14 @@
+package com.easypost.exception.General;
+
+import com.easypost.exception.EasyPostException;
+
+public class InvalidObjectError extends EasyPostException {
+    /**
+     * InvalidObjectError constructor.
+     *
+     * @param message the exception message
+     */
+    public InvalidObjectError(final String message) {
+        super(message);
+    }
+}