+++ src/main/java/com/easypost/model/Error.java
@@ -1,10 +1,67 @@
 package com.easypost.model;
 
+import java.util.List;
+
 public final class Error {
-    private String field;
     private String message;
-    private String suggestion;
     private String code;
+    private List<Error> errors;
+    private String suggestion;
+    private String field;
+
+    /**
+     * Get the field of the error.
+     *
+     * @return field
+     */
+    public String getField() {
+        return field;
+    }
+
+    /**
+     * Set the field of the error.
+     *
+     * @param field field
+     */
+    public void setField(final String field) {
+        this.field = field;
+    }
+
+    /**
+     * Get the suggestion of the error.
+     *
+     * @return suggestion
+     */
+    public String getSuggestion() {
+        return suggestion;
+    }
+
+    /**
+     * Set the suggestion of the error.
+     *
+     * @param suggestion suggestion
+     */
+    public void setSuggestion(final String suggestion) {
+        this.suggestion = suggestion;
+    }
+
+    /**
+     * Get the errors from an Error object.
+     *
+     * @return errors
+     */
+    public List<Error> getErrors() {
+        return errors;
+    }
+
+    /**
+     * Set the errors for an Error object.
+     * 
+     * @param errors errors
+     */
+    public void setErrors(List<Error> errors) {
+        this.errors = errors;
+    }
 
     /**
      * Get the error code.
@@ -25,24 +82,6 @@ public final class Error {
     }
 
     /**
-     * Get the field that caused the error.
-     *
-     * @return field
-     */
-    public String getField() {
-        return field;
-    }
-
-    /**
-     * Set the field that caused the error.
-     *
-     * @param field field
-     */
-    public void setField(final String field) {
-        this.field = field;
-    }
-
-    /**
      * Get the error message.
      *
      * @return message
@@ -58,23 +97,5 @@ public final class Error {
      */
     public void setMessage(final String message) {
         this.message = message;
-    }
-
-    /**
-     * Get the suggestion for the error.
-     *
-     * @return suggestion
-     */
-    public String getSuggestion() {
-        return suggestion;
-    }
-
-    /**
-     * Set the suggestion for the error.
-     *
-     * @param suggestion suggestion
-     */
-    public void setSuggestion(final String suggestion) {
-        this.suggestion = suggestion;
     }
 }