+++ src/main/java/com/easypost/model/SmartrateAccuracy.java
@@ -1,5 +1,9 @@
 package com.easypost.model;
 
+import com.easypost.exception.Constants;
+import com.easypost.exception.EasyPostException;
+import com.easypost.exception.General.InvalidParameterError;
+
 public enum SmartrateAccuracy {
     Percentile50("percentile_50"),
     Percentile75("percentile_75"),
@@ -34,14 +38,14 @@ public enum SmartrateAccuracy {
      *
      * @param keyName the internal key name
      * @return the enum value
-     * @throws IllegalArgumentException if the key name is not found
+     * @throws EasyPostException if the key name is not found
      */
-    public static SmartrateAccuracy getByKeyName(String keyName) throws IllegalArgumentException {
+    public static SmartrateAccuracy getByKeyName(String keyName) throws EasyPostException {
         for (SmartrateAccuracy smartrateAccuracy : values()) {
             if (smartrateAccuracy.getKeyName().equals(keyName)) {
                 return smartrateAccuracy;
             }
         }
-        throw new IllegalArgumentException("Invalid SmartrateAccuracy key name.");
+        throw new InvalidParameterError(String.format(Constants.INVALID_PARAMETER, "smartrateAccuracy"));
     }
 }