+++ src/main/java/com/easypost/model/TimeInTransit.java
@@ -1,6 +1,8 @@
 package com.easypost.model;
 
+import com.easypost.exception.Constants;
 import com.easypost.exception.EasyPostException;
+import com.easypost.exception.General.InvalidParameterError;
 import com.google.gson.annotations.SerializedName;
 
 public final class TimeInTransit {
@@ -173,7 +175,7 @@ public final class TimeInTransit {
             case "percentile_99":
                 return this.percentile99;
             default:
-                throw new EasyPostException("Invalid percentile value");
+                throw new InvalidParameterError(String.format(Constants.INVALID_PARAMETER, "percentile"));
         }
     }